requirements.txt
sqlalchemy==2.0.21
alembic==1.12.0
psycopg2==2.9.9
python-dotenv==1.0.0
fastapi==0.103.2
uvicorn==0.23.2
email-validator==2.0.0.post2
pydantic-extra-types==2.1.0
phonenumbers==8.13.22
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
libgravatar==1.0.4
fastapi-mail==1.4.1
fastapi-limiter==0.1.5
cloudinary==1.36.0
pydantic-settings==2.0.3
fastapi-jwt-auth==0.5.0